backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pdf2image==1.16.3
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
simplekml==1.3.6
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
